move test_wf_to_graph to new test_visualize.py
